Use the simple lrlex API in the example build files.
Since there's a decent chance that people will copy and paste this
stuff, it doesn't seem like a good idea to show them the more complex
API, especially as it's mostly not hugely useful...

+++ lrpar/examples/calc_ast/src/main.rs
@@ -55,7 +55,7 @@ fn main() {
 }
 
 fn eval(
-    lexer: &dyn NonStreamingLexer<DefaultLexeme<u8>, u8>,
+    lexer: &dyn NonStreamingLexer<DefaultLexeme<u32>, u32>,
     e: Expr,
 ) -> Result<u64, (Span, &'static str)> {
     match e {
